[IT-4485] Add endpoint to show GL account balances
Add a new API endpoint `/balances` to output a CSV of GL account
balances appropriate for sending to FloQast.

During the first week of the month, process the balance for the previous
month, otherwise process the month-to-date balance.

The pre-existing endpoints used the Program segment of the chart of
accounts, but the new endpoint uses the GL segment of the chart of
accounts. Refactor fetching the chart of accounts to support arbitrary
segments.